requirements.txt
google-adk
openai>=1.0.0
python-dotenv>=0.19.0
litellm>=1.60.0
requests>=2.31.0